+++ exporter/opentelemetry-exporter-otlp-proto-grpc/src/opentelemetry/exporter/otlp/proto/grpc/exporter.py
@@ -37,7 +37,7 @@ from grpc import (
     ssl_channel_credentials,
 )
 
-from opentelemetry.proto.common.v1.common_pb2 import AnyValue, KeyValue
+from opentelemetry.proto.common.v1.common_pb2 import AnyValue, ArrayValue, KeyValue
 from opentelemetry.proto.resource.v1.resource_pb2 import Resource
 from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_OTLP_CERTIFICATE,
@@ -81,7 +81,7 @@ def environ_to_compression(environ_key: str) -> Optional[Compression]:
     return _ENVIRON_TO_COMPRESSION[environ_value]
 
 
-def _translate_key_values(key: Text, value: Any) -> KeyValue:
+def _translate_value(value: Any) -> KeyValue:
 
     if isinstance(value, bool):
         any_value = AnyValue(bool_value=value)
@@ -96,8 +96,11 @@ def _translate_key_values(key: Text, value: Any) -> KeyValue:
         any_value = AnyValue(double_value=value)
 
     elif isinstance(value, Sequence):
-        any_value = AnyValue(array_value=value)
+        any_value = AnyValue(
+            array_value=ArrayValue(
+                values=[_translate_value(v) for v in value]))
 
+    # Tracing specs currently does not support Mapping type attributes
     elif isinstance(value, Mapping):
         any_value = AnyValue(kvlist_value=value)
 
@@ -106,7 +109,11 @@ def _translate_key_values(key: Text, value: Any) -> KeyValue:
             "Invalid type {} of value {}".format(type(value), value)
         )
 
-    return KeyValue(key=key, value=any_value)
+    return any_value
+
+
+def _translate_key_values(key: Text, value: Any) -> KeyValue:
+    return KeyValue(key=key, value=_translate_value(value))
 
 
 def get_resource_data(